+++ docs/CHANGELOG.md
@@ -11,6 +11,8 @@ and this repository adheres to [Semantic Versioning](https://semver.org/spec/v2.
 
 ### General
 
+- Add `W3C` as default propagator.
+
 ### Breaking changes
 
 ### Enhancements